Update to SLR impact module #2
* Fully reworked the retreat formulation. Now costs of retreat are
tracked as they are accounted for in CIAM, including relocation and
demolition costs.
*  When retreated proactively, assets are not simply lost, but
decomissioned, meaning they are used for another 20 years, while being
depreciated and then only the remaining value after 20 years is assumed
to be a cost of retreat.
* Also added a vulnerability framework to account for vulnerable assets
and people being damaged or retreating first, meaning that the remaining
assets are those that are less vulnerable.
* Changed damage function parameters to match CIAM results in tuning of
globally aggregated setup.